+++ controllers/UserController.js
@@ -1,15 +1,6 @@
 const UserService = require("../services/UserService");
 const errorGenerator = require("../utils/errorGenerator");
 
-const getAddress = async (req, res, next) => {
-  try {
-    const address = await UserService.getAddress();
-    return res.status(200).json({ message: "SUCCESS", address });
-  } catch (err) {
-    res.status(err.statusCode || 500).json({ message: err.message });
-  }
-};
-
 const sendLogIn = async (req, res) => {
   try {
     const { email, password } = req.body;
@@ -41,4 +32,4 @@ const signup = async (req, res) => {
   }
 };
 
-module.exports = { getAddress, sendLogIn, signup };
+module.exports = { sendLogIn, signup };
