+++ routes/index.js
@@ -5,6 +5,7 @@ const UserRoute = require("./UserRoute");
 const CategoryRoute = require("./CategoryRoute");
 const FormRoute = require("./FormRoute");
 const MasterRoute = require("./MasterRoute");
+const AddressRoute = require("./AddressRoute");
 
 router.get("/", (req, res) =>
   res.status(200).json({ message: "Hello! You are connected." })
@@ -14,5 +15,6 @@ router.use("/users", UserRoute);
 router.use("/category", CategoryRoute);
 router.use("/form", FormRoute);
 router.use("/master", MasterRoute);
+router.use("/address", AddressRoute);
 
 module.exports = router;
