+++ services/AddressService.js
@@ -1,0 +1,8 @@
+const AddressDao = require("../models/AddressDao");
+// const errorGenerator = require("../utils/errorGenerator");
+
+const getAddress = async () => {
+  return await AddressDao.getAddress();
+};
+
+module.exports = { getAddress };
